split and streamline hashupdater hook per mapframework remove not needed 2d 3d awareness if any thing the onChange callbacks should become refs to not trigger rerenders on their own.
